# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TripHarmony API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/trip": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Create a new trip",
                "operationId": "6a4632b24da82a47cafb6a2eed78bd33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin",
                                    "destination",
                                    "destination_name"
                                ],
                                "properties": {
                                    "origin": {
                                        "properties": {
                                            "lat": {
                                                "type": "number",
                                                "example": 37.7749
                                            },
                                            "lng": {
                                                "type": "number",
                                                "example": -122.4194
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "destination": {
                                        "properties": {
                                            "lat": {
                                                "type": "number",
                                                "example": 37.7749
                                            },
                                            "lng": {
                                                "type": "number",
                                                "example": -122.4194
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "destination_name": {
                                        "type": "string",
                                        "example": "San Francisco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The new trip",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "origin": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "destination": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "destination_name": {
                                            "type": "string"
                                        },
                                        "driver_id": {
                                            "type": "integer",
                                            "example": null
                                        },
                                        "driver_location": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object",
                                            "example": null
                                        },
                                        "is_started": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "is_complete": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trip/{trip}": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get a trip",
                "operationId": "42f69025753d2723c9c17d46dda66878",
                "parameters": [
                    {
                        "name": "trip",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The trip",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "origin": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "destination": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "destination_name": {
                                            "type": "string"
                                        },
                                        "driver_id": {
                                            "type": "integer",
                                            "example": null
                                        },
                                        "driver_location": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object",
                                            "example": null
                                        },
                                        "is_started": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "is_complete": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find this trip.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot find this trip."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/trip/{trip}/start": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Start a trip",
                "operationId": "96b786bba3e7666f6332028431290120",
                "parameters": [
                    {
                        "name": "trip",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The trip",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "origin": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "destination": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "destination_name": {
                                            "type": "string"
                                        },
                                        "driver_id": {
                                            "type": "integer",
                                            "example": null
                                        },
                                        "driver_location": {
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object",
                                            "example": null
                                        },
                                        "is_started": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_complete": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}